Refactor the libc

+++ plat/cpm/libsys/write.c
@@ -22,7 +22,7 @@ void _sys_write_tty(char c)
 	}
 }
 
-int write(int fd, void* buffer, size_t count)
+ssize_t write(int fd, void* buffer, size_t count)
 {
 	int i;
 	char* p = buffer;
